add unit test for write row field with MySQL number type

+++ core/src/main/java/com/ctrip/framework/drc/core/driver/binlog/impl/AbstractRowsEvent.java
@@ -498,7 +498,7 @@ public abstract class AbstractRowsEvent extends AbstractLogEvent implements Rows
                 if (nbits > 1) {
                     switch (byteCount) {
                         case 1:
-                            ByteHelper.writeUnsignedByte((int) value, out);
+                            ByteHelper.writeUnsignedByte((short) value, out);
                             return;
                         case 2:
                             ByteHelper.writeUnsignedShortBigEndian((int) value, out);
